+++ packages/expo-router/src/imperative-api.ts
@@ -1,6 +1,6 @@
 import { store } from './global-state/router-store';
 import { NavigationOptions } from './global-state/routing';
-import { Href, Route, RouteInputParams } from './types';
+import { Href, Route, RouteInputParams, UnknownInputParams } from './types';
 
 /**
  * Returns `router` object for imperative navigation API.
@@ -73,6 +73,10 @@ export type Router = {
    * @hidden
    */
   reload: () => void;
+  /**
+   * The preload a screen in the background before navigating to it
+   */
+  preload: (name: Href) => void;
 };
 
 /**
@@ -90,4 +94,5 @@ export const router: Router = {
   canGoBack: () => store.canGoBack(),
   setParams: (params) => store.setParams(params),
   reload: () => store.reload(),
+  preload: (name) => store.preload(name),
 };