+++ packages/expo-router/src/link/Link.tsx
@@ -10,6 +10,7 @@ import { useRouter } from '../hooks';
 import { Href } from '../types';
 import { useFocusEffect } from '../useFocusEffect';
 import { useInteropClassName, useHrefAttrs, LinkProps, WebAnchorProps } from './useLinkHooks';
+import { Preload } from '../Preload';
 import { Slot } from '../ui/Slot';
 
 export interface LinkComponent {
@@ -132,6 +133,7 @@ function ExpoRouterLink(
     download,
     withAnchor,
     dangerouslySingular: singular,
+    preload,
     ...rest
   }: LinkProps,
   ref: ForwardedRef<Text>
@@ -169,11 +171,11 @@ function ExpoRouterLink(
     props.onPress(e);
   };
 
-  const Element = asChild ? Slot : Text;
+  const Component = asChild ? Slot : Text;
 
   // Avoid using createElement directly, favoring JSX, to allow tools like NativeWind to perform custom JSX handling on native.
-  return (
-    <Element
+  const element = (
+    <Component
       ref={ref}
       {...props}
       {...hrefAttrs}
@@ -186,6 +188,15 @@ function ExpoRouterLink(
         default: { onPress },
       })}
     />
+  );
+
+  return preload ? (
+    <>
+      <Preload href={href} />
+      {element}
+    </>
+  ) : (
+    element
   );
 }