+++ packages/expo-router/src/link/__tests__/Link.test.ios.tsx
@@ -1,3 +1,4 @@
+import { ParamListBase, StackNavigationState } from '@react-navigation/native';
 import { act, fireEvent, render } from '@testing-library/react-native';
 import React from 'react';
 import { Platform, Text, View } from 'react-native';
@@ -487,5 +488,39 @@ describe('singular', () => {
       stale: false,
       type: 'stack',
     });
+  });
+});
+
+describe('preload', () => {
+  it('can preload the href', () => {
+    renderRouter({
+      index: () => {
+        return <Link preload href="/test" />;
+      },
+      test: () => null,
+    });
+
+    expect(screen).toHaveRouterState({
+      index: 0,
+      key: expect.any(String),
+      preloadedRoutes: [
+        {
+          key: expect.any(String),
+          name: 'test',
+          params: {},
+        },
+      ],
+      routeNames: ['index', 'test', '_sitemap', '+not-found'],
+      routes: [
+        {
+          key: expect.any(String),
+          name: 'index',
+          params: undefined,
+          path: '/',
+        },
+      ],
+      stale: false,
+      type: 'stack',
+    } as StackNavigationState<ParamListBase>);
   });
 });