+++ packages/expo-router/src/link/useLinkHooks.ts
@@ -219,6 +219,10 @@ export interface LinkProps extends Omit<TextProps, 'href'>, WebAnchorProps {
    * If used with `push`, the history will be filtered even if no navigation occurs.
    */
   dangerouslySingular?: SingularOptions;
+  /**
+   * Preloads the route when the component is rendered on a focused screen.
+   */
+  preload?: boolean;
 }
 
 // Mutate the style prop to add the className on web.