+++ packages/expo-router/CHANGELOG.md
@@ -38,6 +38,7 @@ _This version does not introduce any user-facing changes._
 ### 🎉 New features
 
 - Add static redirects to config plugin. ([#34734](https://github.com/expo/expo/pull/34734) by [@marklawlor](https://github.com/marklawlor))
+- Add `router.preload(<href>)` and `<Link preload href={}>`
 
 ### 🐛 Bug fixes